Introduce fine-grained locking and sending of interested / not-interested messages.
There are some debugging assertions added, to ensure we don't send oversized data
packets. These will be removed shortly, but there were crucial in analysing the
sendMessage bug, detailed below.

This commit also fixes a crucial bug in Peer::sendMessage(): if a payload is given,
the message was cut into two: a 5 byte header and the payload itself. However, a
second thread (the uploader) could chime in right between sending the header
and payload and send a piece. The result is that hell would break loose...

The result is that instead of 2MB/sec, upload speeds are improved to ~30MB/sec
locally! The exact cause is not known to me, but alignment may be involved [1]

[1] Discussed with: dwight