Massively restructure the Sender queue and SenderRequest logic.
The problem was that requests would be deleted at any possible time; this
formerly wasn't much of a problem since we would only look at the top of
the queue during the main send loop. However, things got problemetic when
requests would be deleted, as it would be possible that we were servicing
that exact request!

In order to prevent this, transform the queue mutex to a rwlock and
implement cancelling of requests; if a request is cancelled, this may
imply that the peer making the request isn't available anymore and we
should not rely on it.

XXX I think there may be a tiny race when we are about to cancel a request
which has just been handled; I am unsure how to deal with this for now XXX