Fix a serious mistake during handshaking, causing outbound connections to usually fail.
The problem is that all peer activity used the Sender, *except* sending the handshake. This generally wouldn't be a problem, except that the switch to non-blocking sockets resulted in our handshake more-often-than-not being lost.

However. the bitfield message would be queued, which resulted in this message being the first the peer sees. The result is that the peer usually believes we are doing header encryption, etc...

This commit also adds tracing of bad handshakes, and avoids sending pieces to peers of which we know they aren't fully connected yet