Fix a bug which resulted in uploading going completely wrong.
The problem is that the 'last send complete'-flag should only be set after a chunk has been
transferred, not when we are about to transfer one. The end result is that after rescheduling an event, a new event could have been picked because the flag was incorrectly altered, causing data to be interleaved and causing other clients to reject us.