Complete rewrite sending of messages - each Peer is responsible for its messages now.
This is done so we can just assemble a list of peers that need to send data, and sleep until one of their socket becomes writable. The end result is that the 100% CPU use phenomenom is no longer there!

Proper locking of SenderRequest messages is also removed by this commit, since it's no longer relevant: a Peer knows when it's sending, and a message picked for sending will always be removed from the queue and reinserted if needed. This greatly simplifies the locking logic.