Fix a serious bug resulting in corrupted torrents when downloading torrents >4GB on a 32 bit platform.
The problem was the incorrect use of 'size_t' to mark an offset in a file; size_t consists of the maximum length a pointer can point to, which is 4GB on 32 bit platforms (which is why the bug didn't occur on 64 bit platforms). This commit resolves the problem by using 'off_t' when specifying file offsets, which is always >=64 bit (on modern platforms at least) and thus will always yield correct results.