Fix an incredible oversight which would certainly result in random crashes.
The problem was that both the sender and receiver threads could delete a peer object. However, if this happened in the sender, it would not inform the torrent, the sender etc, causing them to use dangling pointers... eventually resulting in massive mayhem.

Why this was not caught before is a mistery to me.