Construct the tracker talker before hashing pieces
As constructing the tracker talker may fail (due to missing/corrupt metadata), we must do so before scheduling pieces since an exception will cause our object construction to be aborted - yet the hasher won't know about this and invoke callbacks in an object already gone.

Note that the tracker won't actually be called during initialization (Torrent::heartbeat() talks to it) so this change does not introduce races.